Fixed issue with creditor amount not being updated properly. Changed doubles to decimals for precision calculation. Created a MoneyDebtReport to return back more information.

+++ TripCalculator/TripCalculator/Models/MoneyTracker.cs
@@ -2,13 +2,13 @@
 {
     public class MoneyTracker
     {
-        public MoneyTracker(Student student, double amount)
+        public MoneyTracker(Student student, decimal amount)
         {
             Student = student;
             Amount = amount;
         }
 
         public Student Student { get; set; }
-        public double Amount { get; set; }
+        public decimal Amount { get; set; }
     }
 }
